+++ e2e/fixtures/plugin/rspress.config.ts
@@ -1,12 +1,10 @@
 import path from 'path';
 import { defineConfig } from 'rspress/config';
-import { pluginPreview } from '@rspress/plugin-preview';
 import { docPluginDemo } from './plugin';
 
 export default defineConfig({
   root: path.join(__dirname, 'doc'),
   plugins: [
     docPluginDemo(),
-    pluginPreview({ isMobile: true, iframePosition: 'follow' }),
   ],
 });