+++ packages/core/src/node/utils/index.ts
@@ -1,5 +1,4 @@
 export * from './normalizePath';
 export * from './getPageKey';
 export * from './createHash';
-export * from './logger';
 export * from './detectReactVersion';