+++ packages/core/src/theme-default/components/Overview/index.tsx
@@ -3,7 +3,12 @@ import { Header, NormalizedSidebarGroup, SidebarItem } from '@rspress/shared';
 import { useSidebarData } from '../../logic';
 import { Link } from '../Link';
 import styles from './index.module.scss';
-import { usePageData, normalizeHref, withBase, isEqualPath } from '@/runtime';
+import {
+  usePageData,
+  normalizeHrefInRuntime as normalizeHref,
+  withBase,
+  isEqualPath,
+} from '@/runtime';
 
 interface GroupItem {
   text?: string;