+++ packages/plugin-last-updated/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @rspress/plugin-last-updated
 
+## 1.0.0-beta.3
+
+### Patch Changes
+
+- rspress@1.0.0-beta.3
+
+## 1.0.0-beta.2
+
+### Patch Changes
+
+- rspress@1.0.0-beta.2
+
+## 1.0.0-beta.1
+
+### Patch Changes
+
+- rspress@1.0.0-beta.1
+
+## 1.0.0-beta.0
+
+### Patch Changes
+
+- rspress@1.0.0-beta.0
+
 ## 0.2.0
 
 ### Minor Changes