+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @rspress/shared
 
+## 1.0.0-beta.3
+
+### Minor Changes
+
+- d1098cf: feat: support cleanUrls to generate url without suffix
+
+### Patch Changes
+
+- a9271e9: feat: using rslog as logger
+
+  feat: 使用 rslog 作为 logger
+
+## 1.0.0-beta.2
+
+### Minor Changes
+
+- d1098cf: feat: support cleanUrls to generate url without suffix
+
+### Patch Changes
+
+- a9271e9: feat: using rslog as logger
+
+  feat: 使用 rslog 作为 logger
+
+## 1.0.0-beta.1
+
+### Minor Changes
+
+- d1098cf: feat: support cleanUrls to generate url without suffix
+
+### Patch Changes
+
+- a9271e9: feat: using rslog as logger
+
+  feat: 使用 rslog 作为 logger
+
+## 1.0.0-beta.0
+
+### Minor Changes
+
+- d1098cf: feat: support cleanUrls to generate url without suffix
+
 ## 0.2.0
 
 ### Minor Changes